requirements.txt
flask
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-pymongo
opentelemetry-exporter-jaeger
pymongo
dnspython
